Addressing PR comments.
(i) Factorizing the unit tests to a separate location  (ii) Readme file (iii) Fix command line message (iv) comments for tests (v) negative test for source_code_unit.get_edit (v) formatting  (vi) updated piranha version v0.1.1  (vii) Renaming and typos.
