+++ generic/piranha/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["Ameya Ketkar <ketkara@uber.com>", "Murali Krishna Ramanathan <murali@uber.com>"]
 name = "piranha"
-version = "2.0.0"
+version = "0.1.1"
 edition = "2021"
 include = [
     "build.rs",
