+++ generic/piranha/src/models/piranha_arguments.rs
@@ -83,25 +83,16 @@ impl PiranhaArguments {
   }
 }
 
-mod test {
-
-  #[cfg(test)]
-  use {
-    super::PiranhaArguments, crate::utilities::tree_sitter_utilities::TreeSitterHelpers,
-    std::collections::HashMap,
-  };
-
-  #[cfg(test)]
-  impl PiranhaArguments {
-    pub(crate) fn dummy() -> Self {
-      let language_name = String::from("java");
-      PiranhaArguments {
-        path_to_code_base: String::new(),
-        input_substitutions: HashMap::new(),
-        path_to_configurations: String::new(),
-        language: language_name.get_language(),
-        language_name,
-      }
+#[cfg(test)]
+impl PiranhaArguments {
+  pub(crate) fn dummy() -> Self {
+    let language_name = String::from("java");
+    PiranhaArguments {
+      path_to_code_base: String::new(),
+      input_substitutions: HashMap::new(),
+      path_to_configurations: String::new(),
+      language: language_name.get_language(),
+      language_name,
     }
   }
 }
