+++ generic/piranha/src/utilities/mod.rs
@@ -111,61 +111,5 @@ pub(crate) fn find_file(input_dir: &PathBuf, name: &str) -> PathBuf {
 }
 
 #[cfg(test)]
-mod test {
-
-  use crate::utilities::find_file;
-  use serde_derive::Deserialize;
-  use std::path::PathBuf;
-
-  #[cfg(test)]
-  use super::{read_file, read_toml};
-
-  #[derive(Deserialize, Default)]
-  struct TestStruct {
-    ip: String,
-  }
-
-  #[test]
-  pub fn test_read_file() {
-    let project_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    let path_to_test_file = project_root.join("test-resources/utility_tests/sample.toml");
-    let result = read_file(&path_to_test_file);
-    assert!(result.is_ok());
-    let content = result.ok().unwrap();
-    assert!(!content.is_empty());
-    assert!(content.eq(r#"ip = '127.0.0.1'"#));
-  }
-
-  #[test]
-  pub fn test_read_toml() {
-    let project_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    let path_to_test_file = project_root.join("test-resources/utility_tests/sample.toml");
-    let result: TestStruct = read_toml(&path_to_test_file, false);
-    assert!(result.ip.eq("127.0.0.1"));
-  }
-
-  #[test]
-  pub fn test_read_toml_default() {
-    let project_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    let path_to_test_file = project_root.join("test-resources/utility_tests/sample1.toml");
-    let result: TestStruct = read_toml(&path_to_test_file, true);
-    assert!(result.ip.eq(""));
-  }
-
-  #[test]
-  pub fn test_find_file_positive() {
-    let project_root =
-      PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-resources/utility_tests/");
-    let f = find_file(&project_root, "sample.toml");
-    assert!(f.is_file());
-  }
-
-  #[test]
-  #[should_panic]
-  pub fn test_find_file_negative() {
-    let project_root =
-      PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test-resources/utility_tests/");
-    let f = find_file(&project_root, "sample1.toml");
-    assert!(f.is_file());
-  }
-}
+#[path = "unit_tests/utilities_test.rs"]
+mod utilities_test;
