+++ generic/piranha/test-resources/utility_tests/sample.toml
@@ -1,1 +1,1 @@
-ip = '127.0.0.1'
+name = 'Piranha'
